Move program types to type folder
