+++ server/graphql/programs/src/mutations/insert_document_registry.rs
@@ -1,4 +1,5 @@
 use async_graphql::*;
+use graphql_types::types::document_registry::{DocumentRegistryNode, DocumentRegistryTypeNode};
 use service::{
     auth::{CapabilityTag, Resource, ResourceAccessRequest},
     document::document_registry::{InsertDocRegistryError, InsertDocumentRegistry},
@@ -8,8 +9,6 @@ use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
-
-use crate::types::document_registry::{DocumentRegistryNode, DocumentRegistryTypeNode};
 
 #[derive(InputObject)]
 pub struct InsertDocumentRegistryInput {
