+++ server/graphql/programs/src/mutations/patient/insert.rs
@@ -3,12 +3,11 @@ use graphql_core::{
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
+use graphql_types::types::patient::PatientNode;
 use service::{
     auth::{CapabilityTag, Resource, ResourceAccessRequest},
     programs::patient::{UpdatePatient, UpdatePatientError},
 };
-
-use crate::types::patient::PatientNode;
 
 #[derive(InputObject)]
 pub struct InsertPatientInput {
