+++ server/graphql/programs/src/queries/document_registry.rs
@@ -1,51 +1,15 @@
 use async_graphql::*;
 use graphql_core::{
-    generic_filters::EqualFilterStringInput,
-    map_filter,
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
-use repository::{
-    DocumentRegistryFilter, DocumentRegistrySort, DocumentRegistrySortField, EqualFilter,
+use graphql_types::types::document_registry::{
+    DocumentRegistryConnector, DocumentRegistryFilterInput, DocumentRegistryNode,
+    DocumentRegistrySortInput,
 };
+use repository::DocumentRegistryFilter;
 use service::auth::{CapabilityTag, Resource, ResourceAccessRequest};
 use service::usize_to_u32;
-
-use crate::types::document_registry::{
-    DocumentRegistryConnector, DocumentRegistryNode, DocumentRegistryTypeNode,
-};
-
-#[derive(InputObject, Clone)]
-pub struct EqualFilterDocumentRegistryTypeInput {
-    pub equal_to: Option<DocumentRegistryTypeNode>,
-    pub equal_any: Option<Vec<DocumentRegistryTypeNode>>,
-    pub not_equal_to: Option<DocumentRegistryTypeNode>,
-}
-
-#[derive(InputObject, Clone)]
-pub struct DocumentRegistryFilterInput {
-    pub id: Option<EqualFilterStringInput>,
-    pub r#type: Option<EqualFilterDocumentRegistryTypeInput>,
-    pub document_type: Option<EqualFilterStringInput>,
-    pub document_context: Option<EqualFilterStringInput>,
-    pub parent_id: Option<EqualFilterStringInput>,
-}
-
-#[derive(Enum, Copy, Clone, PartialEq, Eq)]
-#[graphql(rename_items = "camelCase")]
-pub enum DocumentRegistrySortFieldInput {
-    Type,
-    DocumentType,
-}
-
-#[derive(InputObject)]
-pub struct DocumentRegistrySortInput {
-    /// Sort query result by `key`
-    key: DocumentRegistrySortFieldInput,
-    /// Sort query result is sorted descending or ascending (if not provided the default is
-    /// ascending)
-    desc: Option<bool>,
-}
 
 #[derive(Union)]
 pub enum DocumentRegistryResponse {
@@ -99,37 +63,4 @@ pub fn document_registries(
                 .collect(),
         },
     ))
-}
-
-impl DocumentRegistryFilterInput {
-    pub fn to_domain(self) -> DocumentRegistryFilter {
-        let DocumentRegistryFilterInput {
-            id,
-            r#type,
-            document_type,
-            document_context,
-            parent_id,
-        } = self;
-        DocumentRegistryFilter {
-            id: id.map(EqualFilter::from),
-            document_type: document_type.map(EqualFilter::from),
-            document_context: document_context.map(EqualFilter::from),
-            r#type: r#type.map(|t| map_filter!(t, DocumentRegistryTypeNode::to_domain)),
-            parent_id: parent_id.map(EqualFilter::from),
-        }
-    }
-}
-
-impl DocumentRegistrySortInput {
-    pub fn to_domain(self) -> DocumentRegistrySort {
-        let key = match self.key {
-            DocumentRegistrySortFieldInput::Type => DocumentRegistrySortField::Type,
-            DocumentRegistrySortFieldInput::DocumentType => DocumentRegistrySortField::DocumentType,
-        };
-
-        DocumentRegistrySort {
-            key,
-            desc: self.desc,
-        }
-    }
 }
