--- server/graphql/programs/src/types/mod.rs
@@ -1,6 +1,0 @@
-pub(crate) mod document;
-pub(crate) mod document_registry;
-pub(crate) mod encounter;
-pub(crate) mod patient;
-pub(crate) mod program_enrolment;
-pub(crate) mod program_event;
