+++ server/graphql/types/src/types/program/document.rs
@@ -2,15 +2,91 @@ use async_graphql::*;
 use async_graphql::{dataloader::DataLoader, Context};
 use chrono::{DateTime, Utc};
 
+use graphql_core::generic_filters::{
+    DatetimeFilterInput, EqualFilterStringInput, StringFilterInput,
+};
 use graphql_core::loader::{
     DocumentRegistryLoader, DocumentRegistryLoaderInput, JsonSchemaLoader, UserLoader,
 };
 use graphql_core::{standard_graphql_error::StandardGraphqlError, ContextExt};
-use graphql_types::types::{JSONSchemaNode, UserNode};
-use repository::{unknown_user, Document};
+use repository::{
+    unknown_user, DatetimeFilter, Document, DocumentFilter, DocumentSort, DocumentSortField,
+    EqualFilter, StringFilter,
+};
 use service::document::raw_document::RawDocument;
 
+use crate::types::{JSONSchemaNode, UserNode};
+
 use super::document_registry::DocumentRegistryNode;
+
+#[derive(InputObject, Clone)]
+pub struct DocumentFilterInput {
+    pub name: Option<EqualFilterStringInput>,
+    pub r#type: Option<EqualFilterStringInput>,
+    pub datetime: Option<DatetimeFilterInput>,
+    pub owner: Option<EqualFilterStringInput>,
+    pub context: Option<EqualFilterStringInput>,
+    /// This filter makes it possible to search the raw text json data.
+    /// Be beware of potential performance issues.
+    pub data: Option<StringFilterInput>,
+}
+
+#[derive(Enum, Copy, Clone, PartialEq, Eq)]
+#[graphql(rename_items = "camelCase")]
+pub enum DocumentSortFieldInput {
+    Name,
+    Type,
+    Owner,
+    Context,
+    Datetime,
+}
+
+#[derive(InputObject)]
+pub struct DocumentSortInput {
+    /// Sort query result by `key`
+    key: DocumentSortFieldInput,
+    /// Sort query result is sorted descending or ascending (if not provided the default is
+    /// ascending)
+    desc: Option<bool>,
+}
+
+impl DocumentFilterInput {
+    pub fn to_domain_filter(self) -> DocumentFilter {
+        DocumentFilter {
+            name: self.name.map(|f| repository::StringFilter {
+                equal_to: f.equal_to,
+                not_equal_to: f.not_equal_to,
+                equal_any: f.equal_any,
+                not_equal_all: None,
+                like: None,
+                starts_with: None,
+                ends_with: None,
+            }),
+            r#type: self.r#type.map(EqualFilter::from),
+            datetime: self.datetime.map(DatetimeFilter::from),
+            owner: self.owner.map(EqualFilter::from),
+            context: self.context.map(EqualFilter::from),
+            data: self.data.map(StringFilter::from),
+        }
+    }
+}
+
+impl DocumentSortInput {
+    pub fn to_domain(self) -> DocumentSort {
+        let key = match self.key {
+            DocumentSortFieldInput::Name => DocumentSortField::Name,
+            DocumentSortFieldInput::Type => DocumentSortField::Type,
+            DocumentSortFieldInput::Owner => DocumentSortField::Owner,
+            DocumentSortFieldInput::Context => DocumentSortField::Context,
+            DocumentSortFieldInput::Datetime => DocumentSortField::Datetime,
+        };
+
+        DocumentSort {
+            key,
+            desc: self.desc,
+        }
+    }
+}
 
 pub struct DocumentNode {
     pub allowed_ctx: Vec<String>,
