+++ server/graphql/types/src/types/program/mod.rs
@@ -1,0 +1,6 @@
+pub mod document;
+pub mod document_registry;
+pub mod encounter;
+pub mod patient;
+pub mod program_enrolment;
+pub mod program_event;
