+++ server/graphql/types/src/types/program/program_enrolment.rs
@@ -3,18 +3,104 @@ use chrono::{DateTime, Utc};
 use graphql_core::{
     generic_filters::{DatetimeFilterInput, EqualFilterStringInput},
     loader::DocumentLoader,
+    map_filter,
     pagination::PaginationInput,
     standard_graphql_error::StandardGraphqlError,
     ContextExt,
 };
 use repository::{
-    DatetimeFilter, EncounterFilter, EqualFilter, PaginationOption, ProgramEnrolmentRow,
-    ProgramEnrolmentStatus, ProgramEventFilter, ProgramEventSortField, Sort,
+    DatetimeFilter, EncounterFilter, EqualFilter, PaginationOption, ProgramEnrolmentFilter,
+    ProgramEnrolmentRow, ProgramEnrolmentSort, ProgramEnrolmentSortField, ProgramEnrolmentStatus,
+    ProgramEventFilter, ProgramEventSortField, Sort,
 };
 
-use crate::queries::{EncounterConnector, EncounterFilterInput, EncounterSortInput};
+use super::{
+    document::DocumentNode,
+    encounter::{EncounterConnector, EncounterFilterInput, EncounterNode, EncounterSortInput},
+    program_event::ProgramEventNode,
+};
 
-use super::{document::DocumentNode, encounter::EncounterNode, program_event::ProgramEventNode};
+#[derive(Enum, Copy, Clone, PartialEq, Eq)]
+#[graphql(rename_items = "camelCase")]
+pub enum ProgramEnrolmentSortFieldInput {
+    Type,
+    PatientId,
+    EnrolmentDatetime,
+    ProgramEnrolmentId,
+    Status,
+}
+
+#[derive(InputObject)]
+pub struct ProgramEnrolmentSortInput {
+    /// Sort query result by `key`
+    key: ProgramEnrolmentSortFieldInput,
+    /// Sort query result is sorted descending or ascending (if not provided the default is
+    /// ascending)
+    desc: Option<bool>,
+}
+
+impl ProgramEnrolmentSortInput {
+    pub fn to_domain(self) -> ProgramEnrolmentSort {
+        let key = match self.key {
+            ProgramEnrolmentSortFieldInput::Type => ProgramEnrolmentSortField::Type,
+            ProgramEnrolmentSortFieldInput::PatientId => ProgramEnrolmentSortField::PatientId,
+            ProgramEnrolmentSortFieldInput::EnrolmentDatetime => {
+                ProgramEnrolmentSortField::EnrolmentDatetime
+            }
+            ProgramEnrolmentSortFieldInput::ProgramEnrolmentId => {
+                ProgramEnrolmentSortField::ProgramEnrolmentId
+            }
+            ProgramEnrolmentSortFieldInput::Status => ProgramEnrolmentSortField::Status,
+        };
+
+        ProgramEnrolmentSort {
+            key,
+            desc: self.desc,
+        }
+    }
+}
+
+#[derive(InputObject, Clone)]
+pub struct EqualFilterProgramEnrolmentStatusInput {
+    pub equal_to: Option<ProgramEnrolmentNodeStatus>,
+    pub equal_any: Option<Vec<ProgramEnrolmentNodeStatus>>,
+    pub not_equal_to: Option<ProgramEnrolmentNodeStatus>,
+}
+
+#[derive(InputObject, Clone)]
+pub struct ProgramEnrolmentFilterInput {
+    pub patient_id: Option<EqualFilterStringInput>,
+    pub enrolment_datetime: Option<DatetimeFilterInput>,
+    pub program_enrolment_id: Option<EqualFilterStringInput>,
+    pub status: Option<EqualFilterProgramEnrolmentStatusInput>,
+    /// Same as program enrolment document type
+    pub r#type: Option<EqualFilterStringInput>,
+    /// The program name
+    pub program: Option<EqualFilterStringInput>,
+    pub document_name: Option<EqualFilterStringInput>,
+}
+impl ProgramEnrolmentFilterInput {
+    pub fn to_domain_filter(self) -> ProgramEnrolmentFilter {
+        let ProgramEnrolmentFilterInput {
+            patient_id,
+            enrolment_datetime,
+            program_enrolment_id,
+            status,
+            r#type,
+            program,
+            document_name,
+        } = self;
+        ProgramEnrolmentFilter {
+            patient_id: patient_id.map(EqualFilter::from),
+            enrolment_datetime: enrolment_datetime.map(DatetimeFilter::from),
+            program_enrolment_id: program_enrolment_id.map(EqualFilter::from),
+            status: status.map(|s| map_filter!(s, ProgramEnrolmentNodeStatus::to_domain)),
+            document_name: document_name.map(EqualFilter::from),
+            document_type: r#type.map(EqualFilter::from),
+            context: program.map(EqualFilter::from),
+        }
+    }
+}
 
 #[derive(InputObject, Clone)]
 pub struct ProgramEventFilterInput {
